pyproject.toml
[project]
name = "ml-benchmark-spategan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "polz-j", email = "julius.polz@kit.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "omegaconf",
    "pyyaml",
    "tqdm",
    "einops",
    "numpy>=2.3",
    "pandas>=2.3",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "matplotlib>=3.10",
    "xarray>=2025.9.0",
    "cartopy>=0.25.0",
    "shapely>=2.1.1",
    "pyproj>=3.7.2",
    "netcdf4>=1.7.2",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.7",
    "ipykernel>=6.30.1",
    "networkx>=3.3",
    "fsspec>=2024.6.1",
    "torchinfo",
]

[project.scripts]
ml-benchmark-spategan = "ml_benchmark_spategan:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
